fix spurious utils_sse failures by providing a more robust operator[]
I'm not sure whether it was a GCC miscompilation or not. Basically the
resulting code would first load a scalar value for comparison and only
after that actually store the vector to the memory location it just read
from. According to my understanding of __attribute__((may_alias)) this
should not have happened. Now a helper object makes it a little clearer
to the compiler what is going on.