cmake: disable warning C4290 with MSVC
that warning is useless noise. I don't care whether MSVC would have done
what I wanted without this statement. This statement is required by the
C++ standard, so please don't warn about it.

The switch is in CMakeLists.txt instead of VcMacros.cmake because it's
only needed for unittest.h.

Signed-off-by: Matthias Kretz <kretz@kde.org>