unittest: have COMPARE fall back to MEMCOMPARE automatically
If COMPARE fails to find the necessary operator== to do its normal work
it now silently forwards the compare to MEMCOMPARE. This should make it
easier to get started with a new Vc implementation.

Also MEMCOMPARE now allows to use two different types, as long as their
sizeof is equal.

refs #691

Signed-off-by: Matthias Kretz <kretz@kde.org>