Fix out-of-bounds access on denormals array
The denormals array contains 64 entries. The code assumed that V::Size
is a power of 2 and therefore a simple for loop sufficed to read all
values without out-of-bounds access. Since the addition of SimdArray
types to the math tests this assumption is not correct anymore.
Therefore, a new testAllDenormals functions implements the correct
iteration over the denormals array. All test code is modified to use
testAllDenormals instead of iterating over the array itself.

Signed-off-by: Matthias Kretz <kretz@kde.org>