Prefix generated test names with Test
Have the test structs, which are generated by the TEST macros, use a
name with Test prefix. This resolves an issue where the test name was
equal to a function to be called from inside the test.

Also, clean up the macro parameter name: fun_ -> name_, since this is
the name of the test, not necessarily a function.

Signed-off-by: Matthias Kretz <kretz@kde.org>