Modified range-for test to expect InputIterator
After Vector & Mask return ElementReference on operator[] iterators can
not return an lvalue reference on operator* and thus only implement
InputIterator and OutputIterator. Consequently, the unit test should
only test that range-for works for InputIterator.

Moved the test to iterators.cpp from utils.cpp since that is the natural
place for it now.

Refs: gh-115
Signed-off-by: Matthias Kretz <kretz@kde.org>