where: Fix dangling reference to bool
Have where_expression store a const Mask copy unless the template
argument specifies a reference. In principle the reference is only
really interesting for fixed_size ABI masks, but it doesn't hurt in any
case since a temporary will live to the end of the full-expression
involving the where expression.

Signed-off-by: Matthias Kretz <kretz@kde.org>