MSVC: fix long differences and MSVC madness
Since long uses 32 bits on Windows, the implicit conversions behave
differently. Those differences are captured behind the LONG_IS_LLONG
switches.

But more concerning is, that MSVC will fail some of the
operator_is_substitution_failure tests depending on their order.
Apparently MSVC considers way to many operators as candidate functions
for overload resolution, but only after a instantiation was done before
the overload resolution in question. This is clearly crazy, but I don't
care enough... So just reorder the test so that it works.

Signed-off-by: Matthias Kretz <kretz@kde.org>