Guard bool reduction overloads against implicit conversions
The call `popcount(int())` should be ill-formed while `popcount(bool())`
is overloaded to mirror the behavior of reductions on masks. The same is
true for any other type that is convertible to bool. And the other
mask/bool reductions.
This change introduces detail::exact_bool, which is implicitly
convertible from bool, but no other type. Thus, using exact_bool as
function argument the functions are guarded against the unwanted
conversions.

Refs: mattkretz/wg21-papers#7

Signed-off-by: Matthias Kretz <kretz@kde.org>