Tests: Use correct alignment flag
After adding fixed_size types to the load_store test the test crashed.
Alignment checking showed that aligned loads were called with unaligned
pointers. The issue was loads & stores from offsets of V::size(), which
is not a power of 2. Consequently those were not vector_aligned loads
and stores. The test now determines the right overaligned flag to use
for strided addresses.

Signed-off-by: Matthias Kretz <kretz@kde.org>