Only zero-init padding with -fsignaling-nans
Without -fsignaling-nans GCC may assume that IEEE signaling NaNs don't
generate user-visible traps during floating-point operations.

Also fix simd to be is_trivially_copyable, which was broken because the
move ctor of _SimdWrapper was not defaulted and thus the compiler chose
the catch-all ctor - which isn't trivial.

Adjust the test for triviality accordingly.

Signed-off-by: Matthias Kretz <kretz@kde.org>