Remove __simd_reinterpret_cast in favor of __bit_cast
The std::bit_cast is the most relevant solution to the simd reinterpret
casting problem. It's still a hard problem to write portable and generic
code where such casts are used. simd_math has one instance where it
needs to cast a vector mask to integer vector. But after switching to
__bit_cast the implementation is more efficient - avoiding round-trips
to fixed_size masks.