Proposal: enable implicit and explicit casts for simd_mask
For simd_mask there are many cases where missing implicit casts are just
annoying. However, if the conversion is not free (performance), the
conversion ctor is explicit.

Note that these additional implicit conversions lead to ambiguities with
the binary operators of simd_mask. This needs one additional overload
per operator (&, |, ^, ==, and != are still missing).