Rewrite fixed_size reduction to use more wide vector binary ops
After extended ABIs got the ability to hold non-power-of-2 number of
elements, the old algorithm is not the most efficient approach anymore.
Now, instead blend a partial vector with a full vector, apply the binary
op, and overwrite the invalid results with the values from the initial
full vector. Thus go from {4, 4, 3} via f(4, 4) to {4, 3} and via f(4,
4) to {4} with a final reduction of the remaining vector.