Call _S_masked only on _VecBuiltin
In obscure cases simd_abi::deduce_t could be fixed_size instead of
_VecBuiltin. But either _VecBuiltin::_S_masked is what is needed or the
conditions leading up to this call are wrong. Therefore, first handle
bitmask to vectormask conversions. Then, if the number of elements is
increasing, it is certain that the _VecBuiltin ABI tag provides the
correct _S_masked function to ensure the new elements are zero.

	* experimental/bits/simd_x86.h
	(_MaskImplX86Mixin::_S_to_maskvector): Call
	_VecBuiltin::_S_masked instead of using a deduced ABI tag.